nexus-scripts: Move all variables of configuration to root file
