+++ packages/nexus-scripts/lib/nexus-scripts.js
@@ -1,6 +1,8 @@
+const fs = require('fs');
+const path = require('path');
+
 const { argv } = require('yargs');
 const chalk = require('chalk');
-// const ora = require('ora');
 
 const start = require('./scripts/start');
 
@@ -10,6 +12,34 @@ const commands = {
 
 const commandToRun = commands[argv._[0]];
 
+// TODO: Validate if this directories exists!
+const rootAppDirectoryPath = fs.realpathSync(process.cwd());
+const srcAppDirectoryPath = path.resolve(rootAppDirectoryPath, 'src/');
+const fileAppRootPath = path.resolve(srcAppDirectoryPath, 'index.js');
+const htmlWebpackPluginTemplate = path.resolve(
+  srcAppDirectoryPath,
+  'index.html',
+);
+
+const PORT = argv.port || process.env.PORT || 3000;
+
+const isAnalyzerEnabled = argv.analyzer;
+const { clearConsole, openBrowser } = argv;
+
+const extraConfiguration = {
+  rootAppDirectoryPath,
+  srcAppDirectoryPath,
+  fileAppRootPath,
+  PORT,
+  htmlWebpackPluginTemplate,
+  // extra helper vars
+  isAnalyzerEnabled,
+  clearConsole: clearConsole !== 'false',
+  openBrowser: openBrowser !== 'false',
+};
+
+console.log('running start command', extraConfiguration);
+
 if (!commandToRun) {
   console.error(
     chalk.red(`
@@ -17,5 +47,5 @@ You need to specify a command to run.
   `),
   );
 } else {
-  commandToRun({ argv });
+  commandToRun({ argv, extraConfiguration });
 }
