+++ packages/nexus-scripts/lib/utils/clearConsole.js
