+++ packages/nexus-scripts/lib/webpack/configurations/custom-babel-loader.js
@@ -1,15 +1,11 @@
-const path = require('path');
-
 module.exports = require('babel-loader').custom(() => ({
   config(cfg) {
     return {
       ...cfg.options,
       presets: [
         ...(cfg.options.presets || []),
-        path.resolve(
-          __dirname,
-          '../../../../../babel-preset-webpack-nexus/lib/babel-preset-webpack-nexus.js',
-        ),
+        // eslint-disable-next-line global-require
+        require('babel-preset-webpack-nexus'),
       ],
     };
   },
