+++ packages/nexus-scripts/lib/webpack/configurations/webpack.base.js
