+++ packages/nexus-scripts/lib/webpack/configurations/webpack.prod.js
