+++ packages/nexus-scripts/package.json
@@ -38,6 +38,7 @@
     "@babel/register": "7.4.4",
     "@babel/runtime": "7.4.4",
     "babel-loader": "8.0.5",
+    "babel-preset-webpack-nexus": "^0.3.0",
     "case-sensitive-paths-webpack-plugin": "2.2.0",
     "chalk": "2.4.2",
     "circular-dependency-plugin": "5.0.2",
